SMS-3423: Regenerated G3 files and project tree with latest changes.

+++ apps/g3_apps/g3_modem/firmware/src/config/pic32cx_mtg_ek_pl460_rf215/stack/g3/adaptation/adp_shared_types.h
@@ -146,7 +146,7 @@ typedef struct
     uint16_t datagramTag;
     uint16_t datagramSize;
     bool discoverRoute;
-    uint8_t handle;
+    uintptr_t handle;
     uint8_t maxHops;
     uint8_t dataType;
     uint8_t originalDataType;
